+++ CHANGES.md
@@ -1,8 +1,25 @@
 # Changelog
+
+## 0.2.0 (2024-11-XX)
+
+- Renamed package
+- Complete re-write of functionality
+- Added platform-specific code
+- Added a number of new gestures
+  - Drag and Drop
+  - Zoom In
+  - Zoom Out
+- Added error handling
+- Added API docs
+- Changed to uv for packaging
+
 ## 0.1.3 (2024-09-04)
+
 - Updated import strategy, the structure is now:  
+
     ```python
     from appium.swipe.actions import SwipeActions, SeekDirection, Direction
     ```
+
 - Changed to ruff for linting/formatting
 - Changed to rye for packaging